Create SetupWinRM.ps1
Super Fast Rough Draft to Setup WinRM, I know it's not pretty but it'll do. Will change write-host to write-debug and such later but wanted to get a rough copy in. Fully working in the current state for both slave and master, just not up to my standards for release code. This works on private and public as well as Azure VM's given the right firewall rules are opened outside the OS. Everything inside the OS is done in this. Maybe I'll add that later.